Refactor API to be similar to the CLI
